fix: dinamic client to get mainnet or testnet
The previous UP provider had to be statically configured to operate on mainnet or testnet. Now, it is set up automatically based on the chainId.

This way, I can also set up two different smart contracts to interact with, one for the mainnet and one for the testnet.

benefits: easy to set up for users